test(workflow): W2 V1 PTEH-adapted tests (API, migration, JDBC, route inventory)
- UserWorkflowDefinitionApiTest: real TCP HTTP contract (10 tests, RANDOM_PORT)
- UserWorkflowDefinitionV1SchemaMigrationTest: canonical V1 consolidated schema
  verification (6 tests; active migration == V1, 4 W2 tables, columns/types/nullability,
  edge UNIQUE covers condition_ref, pre-W2 tables preserved)
- UserWorkflowDefinitionJdbcRepositoryTest: JDBC adapter roundtrip/tenant-isolation/
  version-uniqueness/CAS/publish-archive/listByTenant (6 tests, PostgresTestContainerSupport)
- MvcRouteInventoryTest: assert EXPECTED_W2_ROUTE_COUNT = 9 (frozen public surface)

+++ platform-app/src/test/java/com/example/platform/MvcRouteInventoryTest.java
@@ -1,5 +1,7 @@
 package com.example.platform;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 import java.nio.file.*;
 import java.util.*;
 import org.junit.jupiter.api.Test;
@@ -51,7 +53,7 @@ class MvcRouteInventoryTest extends PostgresTestContainerSupport {
                 .filter(p -> p.contains("/workflow-definitions"))
                 .distinct()
                 .count();
-        assertEquals(EXPECTED_W2_ROUTE_COUNT, w2Routes,
+        assertEquals((long) EXPECTED_W2_ROUTE_COUNT, w2Routes,
                 "W2 public route count must be exactly " + EXPECTED_W2_ROUTE_COUNT
                         + " (public-api-contract.tsv); see /tmp/mvc-route-inventory.txt");
     }
